Move last of the functions, but break a test
